fix adding multiple booking products to cart
